+++ sip_router/modules/tm/t_stats.h
@@ -1,6 +1,6 @@
 /*
  *
- * $Id: t_stats.h,v 1.3 2002/09/19 12:23:55 jku Exp $
+ * $Id: t_stats.h,v 1.4 2002/09/24 03:45:57 jiri Exp $
  *
  *
  * Copyright (C) 2001-2003 Fhg Fokus
@@ -36,15 +36,15 @@ extern struct t_stats *cur_stats, *acc_stats;
 
 struct t_stats {
 	/* number of server transactions */
-	unsigned int transactions;
+	unsigned long transactions;
 	/* number of UAC transactions (part of transactions) */
-	unsigned int client_transactions;
+	unsigned long client_transactions;
 	/* number of transactions in wait state */
-	unsigned int waiting;
+	unsigned long waiting;
 	/* number of transactions which completed with this status */
-	unsigned int completed_3xx, completed_4xx, completed_5xx, 
+	unsigned long completed_3xx, completed_4xx, completed_5xx, 
 		completed_6xx, completed_2xx;
-	unsigned int replied_localy;
+	unsigned long replied_localy;
 };
 
 int init_tm_stats(void);
